Small changes and finished BuildAvrLibrary.ps1

+++ Source/Code/ArduinoTemplateLibrary/Array.h
@@ -72,7 +72,7 @@ namespace ATL {
          */
         inline bool IsValidIndex(int16_t index) const
         {
-            return index >= 0 && index < MaxItems;
+            return index >= 0 && index < (int16_t)MaxItems;
         }
 
         /** Finds the index of the first matching item in the array.
@@ -82,7 +82,7 @@ namespace ATL {
          */
         inline int16_t IndexOf(T item) const
         {
-            for (uint16_t i = 0; i < MaxItems; i++)
+            for (int16_t i = 0; i < (int16_t)MaxItems; i++)
             {
                 if (_arr[i] == item)
                     return i;
